# app/static/swagger.yaml
...
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API to manage customers, mechanics, tickets, and inventory for a mechanic shop."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# MARK: Security Definitions
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /tickets:
    post:
      tags: [Tickets]
      summary: "Create ticket"
      description: "Create a new service ticket."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TicketCreate"
      responses:
        201:
          description: "Ticket created"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              customer_id: 1
              ticket_date: "2023-10-01"
              vin: "1HGCM82633A123456"
              mechanics: [1, 2]
              parts: [{ inventory_id: 1, quantity: 2 }]

    get:
      tags: [Tickets]
      summary: "Get all tickets"
      description: "Retrieve all tickets."
      responses:
        200:
          description: "List of tickets"
          schema:
            $ref: "#/definitions/TicketList"
          examples:
            application/json:
              - id: 1
                customer_id: 1
                ticket_date: "2023-10-01"
                vin: "1HGCM82633A123456"
                mechanics: [1, 2]
                parts: [{ inventory_id: 1, quantity: 2 }]
              - id: 2
                customer_id: 2
                ticket_date: "2023-10-02"
                vin: "1HGCM82633A654321"
                mechanics: [3]
                parts: [{ inventory_id: 2, quantity: 1 }]
  /tickets/my-tickets:
    get:
      tags: [Tickets]
      summary: "Get my tickets"
      description: "Retrieve tickets for the logged-in user."
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: "User's tickets"
          schema:
            $ref: "#/definitions/TicketList"
          examples:
            application/json:
                id: 1
                title: "Sample Ticket"
                description: "This is a sample ticket."
  /tickets/{id}:
    get:
      tags: [Tickets]
      summary: "Get specific ticket"
      description: "Retrieve a ticket by ID."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Ticket details"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              title: "Sample Ticket"
              description: "This is a sample ticket."

    patch:
      tags: [Tickets]
      summary: "Update ticket (selected fields)"
      description: "Update selected fields for a ticket."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TicketPatch"
      responses:
        200:
          description: "Ticket updated"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              title: "Sample Ticket"
              description: "This is a sample ticket."

    put:
      tags: [Tickets]
      summary: "Assign/Remove mechanics from ticket"
      description: "Add or remove mechanics assigned to a ticket."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TicketMechanicAssignment"
      responses:
        200:
          description: "Ticket mechanics updated"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              customer_id: 1
              ticket_date: "2023-10-01"
              vin: "1HGCM82633A123456"
              mechanics: [1, 2]
              parts: [{ inventory_id: 1, quantity: 2 }]
  /tickets/{id}/mechanics:
    get:
      tags: [Tickets]
      summary: "Get mechanics assigned to ticket"
      description: "Retrieve all mechanics assigned to a specific ticket."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanics assigned to ticket"
          schema:
            $ref: "#/definitions/MechanicList"
          examples:
            application/json:
              - id: 1
                name: "Mechanic One"
                email: "mechanic.one@example.com"
              - id: 2
                name: "Mechanic Two"
                email: "mechanic.two@example.com"
  /tickets/{id}/assign-mechanic/{mech_id}:
    put:
      tags: [Tickets]
      summary: "Assign mechanic to ticket"
      description: "Assign a mechanic to a ticket."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
        - in: "path"
          name: "mech_id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanic assigned to ticket"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              customer_id: 1
              ticket_date: "2023-10-01"
              vin: "1HGCM82633A123456"
              mechanics: [1, 2]
              parts: [{ inventory_id: 1, quantity: 2 }]
  /tickets/{id}/remove-mechanic/{mech_id}:
    put:
      tags: [Tickets]
      summary: "Remove mechanic from ticket"
      description: "Remove a mechanic from a ticket."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
        - in: "path"
          name: "mech_id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanic removed from ticket"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              customer_id: 1
              ticket_date: "2023-10-01"
              vin: "1HGCM82633A123456"
              mechanics: [1, 2]
              parts: [{ inventory_id: 1, quantity: 2 }]
  /tickets/{id}/add-part:
    post:
      tags: [Tickets]
      summary: "Add inventory item to ticket"
      description: "Add an inventory item to a ticket (part usage)."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TicketPartAdd"
      responses:
        200:
          description: "Inventory item added to ticket"
          schema:
            $ref: "#/definitions/Ticket"
          examples:
            application/json:
              id: 1
              customer_id: 1
              ticket_date: "2023-10-01"
              vin: "1HGCM82633A123456"
              mechanics: [1, 2]
              parts: [{ inventory_id: 1, quantity: 2 }]

  # MARK: Inventory Endpoints
...
definitions:
...
  Mechanic:
    type: "object"
    properties:
      id: { type: "integer" }
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      salary: { type: "number", format: "float" }
  MechanicList:
    type: "array"
    items:
      $ref: "#/definitions/Mechanic"
  TicketCreate:
    type: "object"
    properties:
      customer_id: { type: "integer" }
      ticket_date: { type: "string", format: "date" }
      vin: { type: "string" }
    required: [customer_id, ticket_date, vin]
  TicketPatch:
    type: "object"
    properties:
      ticket_date: { type: "string", format: "date" }
      vin: { type: "string" }
  TicketMechanicAssignment:
    type: "object"
    properties:
      add_mechanic_ids:
        type: "array"
        items: { type: "integer" }
      remove_mechanic_ids:
        type: "array"
        items: { type: "integer" }
  TicketPartAdd:
    type: "object"
    properties:
      inventory_id: { type: "integer" }
      quantity: { type: "integer" }
    required: [inventory_id, quantity]
  Ticket:
    type: "object"
    properties:
      id: { type: "integer" }
      customer_id: { type: "integer" }
      ticket_date: { type: "string" }
      vin: { type: "string" }
      mechanics:
        type: "array"
        items: { type: "integer" }
      parts:
        type: "array"
        items:
          type: "object"
          properties:
            inventory_id: { type: "integer" }
            quantity: { type: "integer" }
  TicketList:
    type: "array"
    items:
      $ref: "#/definitions/Ticket"
...